Parse apache host for registry
In OSP15+, we use an apache vhost to serve the images. This host may be
a hostname instead of an ip address which can lead to failures (404)
when trying to fetch via an address rather than the configured hostname.
This change fetches the apache config and parses out the hostname to use
when testing the pull.

Change-Id: I60f99f5d58ec98112727492c5dc39d3b1cce6724