Improve the endpoint SSL grok
This change updates the endpoint SSL grok so that we're checking
the endpoint list, instead of the service catalog, which is full
of information that the we dont need and is likely to cause a
false positives.

The tasks will now check for the public http endpoint list and if
the list is > 0 it will fail. The failure block will now print
information about the failure so that we're able to debug where
something might have gone wrong in the deployment process.

Signed-off-by: Kevin Carter <kecarter@redhat.com>
Change-Id: I11a062d1b7452f1bb3a2b91340243b44859ce920