Multi cell deployment
Deploy the additonal cell1 with 1 cell controller and
1 compute node.

Create the network configuration for cellcontroller
and add to environment file also as post task validate
the deployment, create the cell and discover compute nodes.

Also added test case to deploy the instance on overcloud(cell0)
and cell1 then delete both instances and execute the
test_archive_database to clean the database for both the cell.

Change-Id: I0a83de413c9f89df5a4ee79c196afca1001be49d