Update multicell deployment to work with tls-e
Infrared with tls-everywhere currently appends environment yamls
public_vip and cloud-names to the overcloud_deployment script. These
interfer with the follow up cell deployment for multi-cellsv2. If
present in the overcloud_cell_deployment.sh script remove them.

post_tasks.sh was also updated to get and apply internal api vip
information from the cell controller and applying it to the control plan
controller. cell1.yaml was also updated to be a template that can read
in the cloudomain of the deployment and apply it accordingly.

These change also allow for a deployment to execute movement based
operations tests, such as live-migration, within cell1.

Change-Id: I8298034baddbc15cb518c128f33615aec31df555