Parse provided OSP version and translate
The cell export osc changes were only updated for 16.2 [1] and not
backported to 16.1. Since IR OSP version translates either the string or
float value to an int this creates an issue when determining which
version of the OSC cell export command should be executed downstream.
This change will use the traditional IR conversion if the provided
version is a string e.g. wallaby, xena, 17-trunk; but will not fully
convert the version if the value provided is already a float e.g. 16.1
and 16.2.

[1] https://review.opendev.org/c/openstack/python-tripleoclient/+/807252

Change-Id: Iec15fee034a4f68badc68902a6e3f2b707a1ac6c