Apply control properties to extra compute
This change is meant to finalize end to end testing for Multi-cellv2 for
17+ RHOS deployments. With the last set of 17 relevant changes coming
via [1]. With this landing the RFE [2] is now in a status where it can
be fully tested in a non-FFU capacity.

In the event that there are no spare controllers when starting the
multi-cell scale up, reassign the extra compute to be a controller
instead. Currently when deploying with extra controllers IR fails during
the intropsection phase. When that is addressed will revert this
approach. Also update user login when using 17 since this has
transitioned from heat-admin to tripleo-admin.

[1] https://bugzilla.redhat.com/show_bug.cgi?id=2089512
[2] https://bugzilla.redhat.com/show_bug.cgi?id=1760256

Change-Id: Ie7471460e91be9714886b129828f2586a4c2dd25